chore(fullstack): tweak bridge example
