+++ examples/fullstack/api/src/resolvers.rs
@@ -3,9 +3,10 @@ use stellation_bridge::links::LocalLink;
 use stellation_bridge::registry::ResolverRegistry;
 use stellation_bridge::resolvers::{MutationResolver, QueryResolver};
 use stellation_bridge::types::{MutationResult, QueryResult};
+use stellation_bridge::Bridge;
 use time::OffsetDateTime;
 
-use crate::types::*;
+pub use crate::types::*;
 
 #[async_trait(?Send)]
 impl QueryResolver for ServerTimeQuery {
@@ -39,3 +40,14 @@ pub fn create_resolver_registry() -> ResolverRegistry<()> {
 }
 
 pub type DefaultLink = LocalLink<()>;
+pub type DefaultBridge = Bridge<DefaultLink>;
+
+pub async fn create_backend_bridge(_req: Option<String>) -> DefaultBridge {
+    Bridge::new(
+        LocalLink::builder()
+            .context(())
+            .resolvers(create_resolver_registry())
+            .routines(create_routine_registry())
+            .build(),
+    )
+}
