+++ examples/fullstack/client/src/main.rs
@@ -3,23 +3,15 @@
 
 mod app;
 use app::App;
-use example_fullstack_api::create_routine_registry;
-use stellation_bridge::links::FetchLink;
-use stellation_bridge::Bridge;
+use example_fullstack_api::create_frontend_bridge;
 use tracing_subscriber::filter::LevelFilter;
 
 fn main() {
+    // Configures Logging
     stellation_frontend::trace::init_default(LevelFilter::INFO);
 
-    // Create bridge
-    let bridge = Bridge::new(
-        FetchLink::builder()
-            .routines(create_routine_registry())
-            .build(),
-    );
-
-    // Start Application
+    // Starts Application
     stellation_frontend::Renderer::<App>::new()
-        .bridge(bridge)
+        .bridge(create_frontend_bridge())
         .render();
 }
